-  make the table more generic in terms of defining what data is contained -  Data is really more defined like rows and columns -  Prop could handle data transformation rather than transformation being built into the table itself -  So basically the table will get the already prepared data in prop  and then just render it
