+++ client/admin/src/store/store.ts
@@ -1,13 +1,13 @@
 import Vue from "vue";
 import Vuex from "vuex";
 import { session } from "./session";
-import { table } from "./table";
+import { products } from "./products";
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
   modules: {
     session,
-    table,
+    products,
   },
 });
